requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121
torch==2.5.1+cu121
torchvision==0.20.1+cu121
transformers>=4.57.0
accelerate
huggingface_hub
pillow
pyyaml
tqdm
numpy
sentencepiece
protobuf
safetensors
